refactor(plot): move legend.py into util directory
