+++ src/caret_analyze/plot/visualize_lib/bokeh/callback_scheduling_source.py
@@ -19,8 +19,7 @@ from typing import Optional
 from bokeh.models import HoverTool
 from bokeh.plotting import ColumnDataSource
 
-from .legend import HoverCreator, HoverKeys, HoverSource, LegendManager
-from .util import RectValues
+from .util import HoverCreator, HoverKeys, HoverSource, LegendManager, RectValues
 
 from ....common import ClockConverter
 from ....record import Clip
