+++ src/caret_analyze/plot/visualize_lib/bokeh/util/__init__.py
@@ -1,5 +1,10 @@
+from .legend import HoverCreator, HoverKeys, HoverSource, LegendManager
 from .util import RectValues
 
 __all__ = [
+    "HoverCreator",
+    "HoverKeys",
+    "HoverSource",
+    "LegendManager",
     "RectValues"
 ]
